Make various normalization tests end-to-end tests
As motivated in commit d8c994d5c08e ("Make BPF symbolization tests
end-to-end tests"), in order to stand a chance of having a reasonable
grasp on what part of the public API surface is being tested, it makes
the most sense to make all end-to-end tests co-located next to each
other (and named consistently etc.).
With this change we take another step into this direction and move all
normalization tests into our end-to-end test suite. To make that happen
we do have to expose a few more types and methods from the crate, but we
make sure to only do so in a hidden manner.
