+++ dev/Cargo.toml
@@ -40,7 +40,10 @@ zip = {version = "2.0.0", optional = true, default-features = false}
 zstd = {version = "0.13.1", default-features = false, optional = true}
 
 [dependencies]
-blazesym = {version = "=0.2.0-rc.2", path = "../"}
+# TODO: Enable `zstd` feature once toolchain support for it is more
+#       widespread (enabled by default in `ld`). Remove conditionals in
+#       test code alongside.
+blazesym = {path = "../", features = ["apk", "breakpad", "gsym", "tracing", "test"]}
 libc = "0.2.137"
 
 [target.'cfg(target_os = "linux")'.dependencies]
