+++ src/normalize/normalizer.rs
@@ -345,323 +345,3 @@ impl Normalizer {
         self.normalize_user_addrs_opts(pid, addrs, &NormalizeOpts::default())
     }
 }
-
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use std::fs::copy;
-    use std::path::Path;
-
-    use tempfile::tempdir;
-
-    use test_log::test;
-
-    use crate::mmap::Mmap;
-    use crate::normalize::buildid::read_elf_build_id;
-    use crate::normalize::Apk;
-    use crate::normalize::Elf;
-    use crate::normalize::Reason;
-    use crate::normalize::Unknown;
-    use crate::normalize::UserMeta;
-    use crate::symbolize;
-    use crate::symbolize::Symbolizer;
-    use crate::test_helper::find_the_answer_fn;
-    use crate::zip;
-
-
-    /// Check that we detect unsorted input addresses.
-    #[test]
-    fn user_address_normalization_unsorted() {
-        let mut addrs = [
-            libc::atexit as Addr,
-            libc::chdir as Addr,
-            libc::fopen as Addr,
-        ];
-        let () = addrs.sort();
-        let () = addrs.swap(0, 1);
-
-        let opts = NormalizeOpts {
-            sorted_addrs: true,
-            ..Default::default()
-        };
-        let normalizer = Normalizer::new();
-        let err = normalizer
-            .normalize_user_addrs_opts(Pid::Slf, addrs.as_slice(), &opts)
-            .unwrap_err();
-        assert!(err.to_string().contains("are not sorted"), "{err}");
-    }
-
-    /// Check that we handle unknown addresses as expected.
-    #[test]
-    fn user_address_normalization_unknown() {
-        // The very first page of the address space should never be
-        // mapped, so use addresses from there.
-        let addrs = [0x500 as Addr, 0x600 as Addr];
-
-        let normalizer = Normalizer::new();
-        let normalized = normalizer
-            .normalize_user_addrs(Pid::Slf, addrs.as_slice())
-            .unwrap();
-        assert_eq!(normalized.outputs.len(), 2);
-        assert_eq!(normalized.meta.len(), 1);
-        assert_eq!(normalized.meta[0], Unknown::new(Reason::Unmapped).into());
-        assert_eq!(normalized.outputs[0].1, 0);
-        assert_eq!(normalized.outputs[1].1, 0);
-    }
-
-    /// Check that we can normalize user addresses.
-    #[cfg(linux)]
-    // `libc` on Arm doesn't have `__errno_location`.
-    #[cfg(not(any(target_arch = "arm", target_arch = "aarch64")))]
-    #[test]
-    fn user_address_normalization() {
-        fn test(normalizer: &Normalizer) {
-            let addrs = [
-                libc::__errno_location as Addr,
-                libc::dlopen as Addr,
-                libc::fopen as Addr,
-                user_address_normalization_unknown as Addr,
-                user_address_normalization as Addr,
-                Mmap::map as Addr,
-            ];
-
-            let (errno_idx, _) = addrs
-                .iter()
-                .enumerate()
-                .find(|(_idx, addr)| **addr == libc::__errno_location as Addr)
-                .unwrap();
-
-            let normalized = normalizer
-                .normalize_user_addrs(Pid::Slf, addrs.as_slice())
-                .unwrap();
-            assert_eq!(normalized.outputs.len(), 6);
-
-            let outputs = &normalized.outputs;
-            let meta = &normalized.meta;
-            assert_eq!(meta.len(), 2);
-
-            let errno_meta_idx = outputs[errno_idx].1;
-            assert!(meta[errno_meta_idx]
-                .as_elf()
-                .unwrap()
-                .path
-                .file_name()
-                .unwrap()
-                .to_string_lossy()
-                .contains("libc.so"));
-        }
-
-        let normalizer = Normalizer::new();
-        test(&normalizer);
-
-        let normalizer = Normalizer::builder().enable_vma_caching(true).build();
-        test(&normalizer);
-        test(&normalizer);
-    }
-
-    /// Check that we can normalize user addresses in our own shared object.
-    #[test]
-    fn user_address_normalization_custom_so() {
-        let test_so = Path::new(&env!("CARGO_MANIFEST_DIR"))
-            .join("data")
-            .join("libtest-so.so");
-
-        let mmap = Mmap::builder().exec().open(&test_so).unwrap();
-        let (sym, the_answer_addr) = find_the_answer_fn(&mmap);
-
-        let normalizer = Normalizer::new();
-        let normalized = normalizer
-            .normalize_user_addrs(Pid::Slf, [the_answer_addr as Addr].as_slice())
-            .unwrap();
-        assert_eq!(normalized.outputs.len(), 1);
-        assert_eq!(normalized.meta.len(), 1);
-
-        let output = normalized.outputs[0];
-        assert_eq!(output.0, sym.file_offset.unwrap());
-        let meta = &normalized.meta[output.1];
-        let expected_elf = Elf {
-            build_id: Some(read_elf_build_id(&test_so).unwrap().unwrap()),
-            path: test_so.clone(),
-            _non_exhaustive: (),
-        };
-        assert_eq!(meta, &UserMeta::Elf(expected_elf));
-    }
-
-    fn test_user_address_normalization_deleted_so(use_procmap_query: bool) {
-        fn test(
-            use_procmap_query: bool,
-            cache_vmas: bool,
-            cache_build_ids: bool,
-            use_map_files: bool,
-        ) {
-            let test_so = Path::new(&env!("CARGO_MANIFEST_DIR"))
-                .join("data")
-                .join("libtest-so.so");
-            let dir = tempdir().unwrap();
-            let tmp_so = dir.path().join("libtest-so.so");
-            let _count = copy(&test_so, &tmp_so).unwrap();
-
-            let mmap = Mmap::builder().exec().open(&tmp_so).unwrap();
-            let (sym, the_answer_addr) = find_the_answer_fn(&mmap);
-
-            // Remove the temporary directory and with it the mapped shared
-            // object.
-            let () = drop(dir);
-
-            let opts = NormalizeOpts {
-                sorted_addrs: false,
-                map_files: use_map_files,
-                ..Default::default()
-            };
-            let normalizer = Normalizer::builder()
-                .enable_procmap_query(use_procmap_query)
-                .enable_vma_caching(cache_vmas)
-                .enable_build_id_caching(cache_build_ids)
-                .build();
-            let normalized = normalizer
-                .normalize_user_addrs_opts(Pid::Slf, [the_answer_addr as Addr].as_slice(), &opts)
-                .unwrap();
-            assert_eq!(normalized.outputs.len(), 1);
-            assert_eq!(normalized.meta.len(), 1);
-
-            let output = normalized.outputs[0];
-            assert_eq!(output.0, sym.file_offset.unwrap());
-            let meta = &normalized.meta[output.1].as_elf().unwrap();
-            let expected_build_id = if use_map_files || use_procmap_query {
-                Some(read_elf_build_id(&test_so).unwrap().unwrap())
-            } else {
-                None
-            };
-
-            assert_eq!(meta.build_id, expected_build_id);
-        }
-
-        for cache_build_ids in [true, false] {
-            for cache_vmas in [true, false] {
-                for use_map_files in [true, false] {
-                    let () = test(
-                        use_procmap_query,
-                        cache_build_ids,
-                        cache_vmas,
-                        use_map_files,
-                    );
-                }
-            }
-        }
-    }
-
-    /// Check that we can normalize user addresses in a shared object
-    /// that has been deleted already (but is still mapped) without
-    /// errors.
-    #[test]
-    fn user_address_normalization_deleted_so_proc_maps() {
-        test_user_address_normalization_deleted_so(false)
-    }
-
-    /// Check that we can normalize user addresses in a shared object
-    /// that has been deleted already (but is still mapped) without
-    /// errors.
-    #[test]
-    #[ignore = "test requires PROCMAP_QUERY ioctl kernel support"]
-    fn user_address_normalization_deleted_so_ioctl() {
-        test_user_address_normalization_deleted_so(true)
-    }
-
-    /// Check that we can normalize addresses in our own shared object inside a
-    /// zip archive.
-    #[test]
-    fn normalize_custom_so_in_zip() {
-        fn test(so_name: &str) {
-            let test_zip = Path::new(&env!("CARGO_MANIFEST_DIR"))
-                .join("data")
-                .join("test.zip");
-
-            let mmap = Mmap::builder().exec().open(&test_zip).unwrap();
-            let archive = zip::Archive::with_mmap(mmap.clone()).unwrap();
-            let so = archive
-                .entries()
-                .find_map(|entry| {
-                    let entry = entry.unwrap();
-                    (entry.path == Path::new(so_name)).then_some(entry)
-                })
-                .unwrap();
-
-            let elf_mmap = mmap
-                .constrain(so.data_offset..so.data_offset + so.data.len() as u64)
-                .unwrap();
-            let (sym, the_answer_addr) = find_the_answer_fn(&elf_mmap);
-
-            let opts = NormalizeOpts {
-                sorted_addrs: true,
-                ..Default::default()
-            };
-            let normalizer = Normalizer::new();
-            let normalized = normalizer
-                .normalize_user_addrs_opts(Pid::Slf, [the_answer_addr as Addr].as_slice(), &opts)
-                .unwrap();
-            assert_eq!(normalized.outputs.len(), 1);
-            assert_eq!(normalized.meta.len(), 1);
-
-            let expected_offset = so.data_offset + sym.file_offset.unwrap();
-            let output = normalized.outputs[0];
-            assert_eq!(output.0, expected_offset);
-            let meta = &normalized.meta[output.1];
-            let expected = Apk {
-                path: test_zip.clone(),
-                _non_exhaustive: (),
-            };
-            assert_eq!(meta, &UserMeta::Apk(expected));
-
-            // Also symbolize the normalization output.
-            let apk = symbolize::Apk::new(test_zip);
-            let src = symbolize::Source::Apk(apk);
-            let symbolizer = Symbolizer::new();
-            let result = symbolizer
-                .symbolize_single(&src, symbolize::Input::FileOffset(output.0))
-                .unwrap()
-                .into_sym()
-                .unwrap();
-
-            assert_eq!(result.name, "the_answer");
-
-            let results = symbolizer
-                .symbolize(&src, symbolize::Input::FileOffset(&[output.0]))
-                .unwrap();
-            assert_eq!(results.len(), 1);
-
-            let sym = results[0].as_sym().unwrap();
-            assert_eq!(sym.name, "the_answer");
-        }
-
-        test("libtest-so.so");
-        test("libtest-so-no-separate-code.so");
-    }
-
-    /// Make sure that when using the `map_files` normalization option,
-    /// we never end up reporting a path referencing "self".
-    #[test]
-    fn normalize_no_self_vma_path_reporting() {
-        let opts = NormalizeOpts {
-            sorted_addrs: true,
-            map_files: true,
-            ..Default::default()
-        };
-        let normalizer = Normalizer::new();
-        let normalized = normalizer
-            .normalize_user_addrs_opts(
-                Pid::Slf,
-                [normalize_no_self_vma_path_reporting as Addr].as_slice(),
-                &opts,
-            )
-            .unwrap();
-
-        assert_eq!(normalized.outputs.len(), 1);
-        assert_eq!(normalized.meta.len(), 1);
-        let output = normalized.outputs[0];
-        let meta = &normalized.meta[output.1];
-        let elf = meta.as_elf().unwrap();
-        assert!(!elf.path.to_string_lossy().contains("self"), "{elf:?}");
-    }
-}
